app/events/requirements.txt
python-dotenv==0.16.0
requests==2.25.1
psycopg2==2.8.6
pymssql==2.1.5
SQLAlchemy==1.4.3
configparser==5.0.2
PyMySQL==1.0.2
pandas==1.2.3
mongoengine==0.23.0
pymongo==3.11.3
numpy==1.20.2
confluent-kafka==1.7.0
faker==13.15.0
faker_airtravel==0.4
Flask==1.1.2
minio==7.1.0
azure-storage-blob==12.8.0
azure-cosmos==4.2.0
azure-eventhub==5.4.0
cassandra-driver==3.25.0
google-cloud-pubsub==2.6.0
confluent_avro==1.8.0
avro-python3==1.10.2
pulsar-client==2.10.0
fastapi==0.78.0
uvicorn==0.18.2